+++ lib/schleuder/subscription.rb
@@ -12,6 +12,8 @@ module Schleuder
 
     default_scope { order(:email) }
 
+    scope :without_fingerprint, -> { where(fingerprint: [nil,'']) }
+
     def to_s
       email
     end